+++ src/main/kotlin/dev/aaronhowser/mods/excessive_utilities/item/ChickenWingRingItem.kt
@@ -101,7 +101,7 @@ class ChickenWingRingItem(properties: Properties) : Item(properties), ICurioItem
 		}
 
 		private fun addGpConsumer(player: ServerPlayer, ringStack: ItemStack): GridPowerContribution.HeldItem {
-			val handler = GridPowerHandler.get(player.serverLevel()).getGrid(player)
+			val handler = GridPowerHandler.get(player.serverLevel()).getPlayerGrid(player)
 
 			val currentConsumers = handler.getConsumers()
 			val existing = currentConsumers